--- app/src/app/utils/api/apiHandlers.ts
@@ -1,33 +1,0 @@
-// Re-export all types from types.ts
-export * from './apiHandlers/types';
-
-// Re-export utility functions
-export { lastSentAssignments, FormatAssignments } from './apiHandlers/formatAssignments';
-
-// Re-export document-related handlers
-export { createMapDocument } from './apiHandlers/createMapDocument';
-export { getDocument } from './apiHandlers/getDocument';
-export { unlockMapDocument } from './apiHandlers/unlockMapDocument';
-export { getMapLockStatus } from './apiHandlers/getMapLockStatus';
-
-// Re-export assignments handlers
-export { getAssignments } from './apiHandlers/getAssignments';
-export { patchUpdateAssignments } from './apiHandlers/patchUpdateAssignments';
-export { patchUpdateReset } from './apiHandlers/patchUpdateReset';
-
-// Re-export contiguity handlers
-export { getContiguity } from './apiHandlers/getContiguity';
-export { getZoneConnectedComponentBBoxes } from './apiHandlers/getZoneConnectedComponentBBoxes';
-
-// Re-export shatter handlers
-export { patchShatterParents } from './apiHandlers/patchShatterParents';
-export { patchUnShatterParents } from './apiHandlers/patchUnShatterParents';
-
-// Re-export other handlers
-export { getAvailableDistrictrMaps } from './apiHandlers/getAvailableDistrictrMaps';
-export { saveMapDocumentMetadata } from './apiHandlers/saveMapDocumentMetadata';
-export { getSharePlanLink } from './apiHandlers/getSharePlanLink';
-export { getLoadPlanFromShare } from './apiHandlers/getLoadPlanFromShare';
-export { checkoutMapDocument } from './apiHandlers/checkoutMapDocument';
-export { saveColorScheme } from './apiHandlers/saveColorScheme';
-export { getDemography } from './apiHandlers/getDemography';
