+++ app/src/app/utils/api/mutations/document.ts
@@ -1,6 +1,6 @@
 import { MutationObserver } from '@tanstack/query-core';
 import { queryClient } from '../queryClient';
-import { createMapDocument } from '../apiHandlers';
+import { createMapDocument } from '../apiHandlers/createMapDocument';
 import { useMapStore } from '@/app/store/mapStore';
 
 export const document = new MutationObserver(queryClient, {
