+++ app/src/app/utils/api/mutations/patchUnShatter.ts
@@ -1,6 +1,6 @@
 import { MutationObserver } from '@tanstack/query-core';
 import { queryClient } from '../queryClient';
-import { patchUnShatterParents } from '../apiHandlers';
+import { patchUnShatterParents } from '../apiHandlers/patchUnShatterParents';
 import { useMapStore } from '@/app/store/mapStore';
 
 export const patchUnShatter = new MutationObserver(queryClient, {
