+++ app/src/app/utils/api/mutations/patchUpdates.ts
@@ -1,6 +1,7 @@
 import { MutationObserver } from '@tanstack/query-core';
 import { queryClient } from '../queryClient';
-import { patchUpdateAssignments, AssignmentsCreate } from '../apiHandlers';
+import { patchUpdateAssignments } from '../apiHandlers/patchUpdateAssignments';
+import {AssignmentsCreate} from '@utils/api/apiHandlers/types';
 import { useMapStore } from '@/app/store/mapStore';
 
 export const patchUpdates = new MutationObserver(queryClient, {
