+++ app/src/app/utils/api/mutations/sharePlan.ts
@@ -1,6 +1,6 @@
 import { MutationObserver } from '@tanstack/query-core';
 import { queryClient } from '../queryClient';
-import { getSharePlanLink } from '../apiHandlers';
+import { getSharePlanLink } from '../apiHandlers/getSharePlanLink';
 import { useMapStore } from '@/app/store/mapStore';
 
 export const sharePlan = new MutationObserver(queryClient, {
